refactor: delegate enemy tick
