+++ debugTools.js
@@ -12,7 +12,6 @@ import {
   startNewGame,
   cDealerDamage
 } from './script.js';
-import { activeDisciples } from './game/disciples.js';
 import {
   stageData,
   timeScale,
